requirements/ci.txt
amqp==2.6.1
    # via kombu
anyascii==0.3.2
    # via wagtail
asgiref==3.6.0
    # via django
beautifulsoup4==4.9.3
    # via wagtail
billiard==3.6.4.0
    # via celery
boto3==1.26.114
    # via newamerica-cms (pyproject.toml)
botocore==1.29.114
    # via
    #   boto3
    #   s3transfer
cairocffi==1.5.0
    # via
    #   cairosvg
    #   weasyprint
cairosvg==2.7.0
    # via weasyprint
celery==4.4.7
    # via newamerica-cms (pyproject.toml)
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via
    #   cairocffi
    #   weasyprint
charset-normalizer==3.1.0
    # via requests
createsend==7.0.0
    # via newamerica-cms (pyproject.toml)
cssselect2==0.7.0
    # via
    #   cairosvg
    #   weasyprint
defusedxml==0.7.1
    # via
    #   cairosvg
    #   willow
dj-database-url==1.3.0
    # via newamerica-cms (pyproject.toml)
django==3.2.18
    # via
    #   dj-database-url
    #   django-anymail
    #   django-appconf
    #   django-basic-auth-ip-whitelist
    #   django-cors-headers
    #   django-csp
    #   django-filter
    #   django-modelcluster
    #   django-multiselectfield
    #   django-permissionedforms
    #   django-redis
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   newamerica-cms (pyproject.toml)
    #   wagtail
django-anymail==9.1
    # via newamerica-cms (pyproject.toml)
django-appconf==1.0.5
    # via django-compressor
django-basic-auth-ip-whitelist==0.5
    # via newamerica-cms (pyproject.toml)
django-compressor==4.3.1
    # via newamerica-cms (pyproject.toml)
django-cors-headers==3.14.0
    # via newamerica-cms (pyproject.toml)
django-csp==3.7
    # via newamerica-cms (pyproject.toml)
django-filter==21.1
    # via
    #   newamerica-cms (pyproject.toml)
    #   wagtail
django-modelcluster==6.0
    # via
    #   newamerica-cms (pyproject.toml)
    #   wagtail
django-multiselectfield==0.1.12
    # via newamerica-cms (pyproject.toml)
django-permissionedforms==0.1
    # via wagtail
django-redis==5.2.0
    # via newamerica-cms (pyproject.toml)
django-storages==1.13.2
    # via newamerica-cms (pyproject.toml)
django-taggit==2.1.0
    # via wagtail
django-treebeard==4.6.1
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
elasticsearch==5.5.3
    # via newamerica-cms (pyproject.toml)
et-xmlfile==1.1.0
    # via openpyxl
faker==18.4.0
    # via newamerica-cms (pyproject.toml)
filetype==1.2.0
    # via willow
gunicorn==20.0.4
    # via newamerica-cms (pyproject.toml)
html5lib==1.1
    # via
    #   wagtail
    #   weasyprint
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==4.6.11
    # via celery
l18n==2021.3
    # via wagtail
lxml==4.9.2
    # via python-docx
openpyxl==3.1.2
    # via wagtail
pillow==9.5.0
    # via
    #   cairosvg
    #   wagtail
psycopg2==2.9.6
    # via newamerica-cms (pyproject.toml)
pycparser==2.21
    # via cffi
pyphen==0.14.0
    # via weasyprint
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
python-docx==0.8.11
    # via newamerica-cms (pyproject.toml)
pytz==2023.3
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
rcssmin==1.1.1
    # via django-compressor
redis==3.5.3
    # via
    #   django-redis
    #   newamerica-cms (pyproject.toml)
requests==2.28.2
    # via
    #   django-anymail
    #   wagtail
rjsmin==1.2.1
    # via django-compressor
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   createsend
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.4
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
telepath==0.3
    # via wagtail
tinycss2==1.2.1
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
typing-extensions==4.5.0
    # via dj-database-url
urllib3==1.26.15
    # via
    #   botocore
    #   elasticsearch
    #   requests
vine==1.3.0
    # via
    #   amqp
    #   celery
wagtail==5.0.5
    # via
    #   newamerica-cms (pyproject.toml)
    #   wagtail-autocomplete
    #   wagtail-headless-preview
wagtail-autocomplete==0.10.0
    # via newamerica-cms (pyproject.toml)
wagtail-headless-preview==0.7.0
    # via newamerica-cms (pyproject.toml)
wand==0.6.11
    # via newamerica-cms (pyproject.toml)
weasyprint==51
    # via newamerica-cms (pyproject.toml)
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise==6.4.0
    # via newamerica-cms (pyproject.toml)
willow==1.6.2
    # via wagtail

# The following packages are considered to be unsafe in a requirements file:
setuptools==67.6.1
    # via
    #   gunicorn
    #   weasyprint
